Introduce `bundle_name` setting
A classic approach to handling bundles is to use a hash of the file content as a filename and attach 'Cache-Control: max-age=31556900' or 'Expires' HTTP header.

Now imagine that a have a website with 100 HTML pages. All of them include js like so:

    <script scr="95a78b02de98.js"></script>

When we change JS, the hash of the file changes too, causing all HTML files that uses that JS to change as well.

One solution to this problem would be to stop using hashes as filenames.